+++ src/backend/token_exchange/tests/test_api.py
@@ -1,18 +1,20 @@
 """Menshen: views tests for the token_exchange application."""
 
+import base64
 import logging
 from datetime import UTC, datetime, timedelta
 from unittest import mock
 from uuid import uuid4
 
 import pytest
+from ninja.testing import TestClient
 from requests import HTTPError
-from rest_framework import status
-from rest_framework.test import APIClient
 
+from token_exchange.api import api
 from token_exchange.enums import TokenType
 from token_exchange.factories import (
     ExchangedTokenFactory,
+    ServiceProviderCredentialsFactory,
     ServiceProviderFactory,
     TokenExchangeRuleFactory,
 )
@@ -20,68 +22,163 @@ from token_exchange.models import ExchangedToken, TokenTypeChoices
 from token_exchange.services.token import TokenGenerator
 
 
+@pytest.mark.parametrize(
+    "endpoint",
+    [
+        "/exchange/",
+        "/introspect/",
+        "/revoke/",
+    ],
+)
 @pytest.mark.django_db
-def test_exchange_view_auth():
-    """Test the TokenExchangeView authentication."""
-    client = APIClient()
-    response = client.post("/auth/token/exchange/", {}, content_type="application/json")
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
+def test_exchange_view_auth_without_credentials(endpoint):
+    """Test the TokenExchangeView authentication without submitting credentials."""
+    client = TestClient(api)
+    response = client.post(endpoint, data={})
+    assert response.status_code == 401  # UNAUTHORIZED
+
+
+@pytest.mark.parametrize(
+    "endpoint",
+    [
+        "/exchange/",
+        "/introspect/",
+        "/revoke/",
+    ],
+)
+@pytest.mark.django_db
+def test_exchange_view_auth_with_invalid_credentials(endpoint):
+    """Test the TokenExchangeView authentication with invalid credentials."""
+    client = TestClient(api)
+    encoded_credentials = base64.b64encode(bytes("foo:bar", encoding="utf-8"))
+    response = client.post(
+        endpoint,
+        headers={"Authorization": "Basic " + encoded_credentials.decode()},
+        data={},
+    )
+    assert response.status_code == 401  # UNAUTHORIZED
+
+
+@pytest.mark.parametrize(
+    "endpoint",
+    [
+        "/exchange/",
+        "/introspect/",
+        "/revoke/",
+    ],
+)
+@pytest.mark.django_db
+def test_exchange_view_auth_with_inactive_service(endpoint):
+    """Test the TokenExchangeView authentication for an inactive service provider."""
+    client = TestClient(api)
+
+    # Create the inactive service
+    service_provider = ServiceProviderFactory.create(audience_id="test:inactive-service")
+    credentials = ServiceProviderCredentialsFactory(
+        service_provider=service_provider, is_active=False
+    )
+
+    encoded_credentials = base64.b64encode(
+        bytes(f"{credentials.client_id}:{credentials.client_secret}", encoding="utf-8")
+    )
+    response = client.post(
+        endpoint,
+        headers={"Authorization": "Basic " + encoded_credentials.decode()},
+        data={},
+    )
+    assert response.status_code == 401  # UNAUTHORIZED
 
 
 def test_exchange_view_invalid_content_type(source_api_client):
     """Test the TokenExchangeView with an invalid request content-type."""
-    response = source_api_client.post("/auth/token/exchange/", data={"token": "invalid"})
-    json_response = response.json()
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert json_response["error"] == "invalid_request"
-    assert json_response["error_description"] == "Request content-type is not JSON."
+    response = source_api_client.post("/exchange/", json={"token": "invalid"})
+    assert response.status_code == 422  # UNPROCESSABLE_ENTITY
+    assert response.json() == {
+        "detail": [
+            {"type": "missing", "loc": ["form", "subject_token"], "msg": "Field required"},
+            {"type": "missing", "loc": ["form", "subject_token_type"], "msg": "Field required"},
+            {"type": "missing", "loc": ["form", "grant_type"], "msg": "Field required"},
+        ]
+    }
 
 
 def test_exchange_view_invalid_payload(source_api_client):
     """Test the TokenExchangeView with a valid content-type but an invalid payload."""
     response = source_api_client.post(
-        "/auth/token/exchange/",
+        "/exchange/",
         data=b"token=invalid",
         content_type="application/json",
     )
-    json_response = response.json()
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert json_response["error"] == "invalid_request"
-    assert json_response["error_description"] == "JSON is malformed: invalid character (byte 4)"
+    assert response.status_code == 422  # UNPROCESSABLE_ENTITY
+    assert response.json() == {
+        "detail": [
+            {"type": "missing", "loc": ["form", "subject_token"], "msg": "Field required"},
+            {"type": "missing", "loc": ["form", "subject_token_type"], "msg": "Field required"},
+            {"type": "missing", "loc": ["form", "grant_type"], "msg": "Field required"},
+        ]
+    }
 
 
 @pytest.mark.django_db
 @pytest.mark.parametrize(
-    "payload",
+    ("payload", "error"),
     [
-        {},
-        {
-            "grant_type": "urn:ietf:params:oauth:grant-type:token-exchange",
-            "subject_token": None,
-            "subject_token_type": "urn:ietf:params:oauth:token-type:access_token",
-            "audience": "service:target",
-        },
-        {
-            "grant_type": "urn:ietf:params:oauth:grant-type:token-exchange",
-            "subject_token": "",
-            "subject_token_type": "urn:ietf:params:oauth:token-type:access_token",
-            "audience": "service:target",
-        },
-        {
-            "grant_type": "urn:ietf:params:oauth:grant-type:token-exchange",
-            "subject_token": "",
-            "subject_token_type": "urn:ietf:params:oauth:token-type:jwt",
-            "audience": "service:target",
-        },
+        (
+            {},
+            {
+                "type": "missing",
+                "loc": ["form", "subject_token"],
+                "msg": "Field required",
+            },
+        ),
+        (
+            {
+                "grant_type": "urn:ietf:params:oauth:grant-type:token-exchange",
+                "subject_token": None,
+                "subject_token_type": "urn:ietf:params:oauth:token-type:access_token",
+                "audience": "service:target",
+            },
+            {
+                "type": "missing",
+                "loc": ["form", "subject_token"],
+                "msg": "Field required",
+            },
+        ),
+        (
+            {
+                "grant_type": "urn:ietf:params:oauth:grant-type:token-exchange",
+                "subject_token": "",
+                "subject_token_type": "urn:ietf:params:oauth:token-type:access_token",
+                "audience": "service:target",
+            },
+            {
+                "type": "string_too_short",
+                "loc": ["form", "subject_token"],
+                "msg": "String should have at least 1 character",
+                "ctx": {"min_length": 1},
+            },
+        ),
+        (
+            {
+                "grant_type": "urn:ietf:params:oauth:grant-type:token-exchange",
+                "subject_token": "",
+                "subject_token_type": "urn:ietf:params:oauth:token-type:jwt",
+                "audience": "service:target",
+            },
+            {
+                "type": "string_too_short",
+                "loc": ["form", "subject_token"],
+                "msg": "String should have at least 1 character",
+                "ctx": {"min_length": 1},
+            },
+        ),
     ],
 )
-def test_exchange_view_invalid_token(source_api_client, payload):
+def test_exchange_view_invalid_token(source_api_client, payload, error):
     """Test the TokenExchangeView with an invalid access/JWT token."""
-    response = source_api_client.post(
-        "/auth/token/exchange/", payload, content_type="application/json"
-    )
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.json()["error"] == "invalid_request"
+    response = source_api_client.post("/exchange/", data=payload)
+    assert response.status_code == 422  # UNPROCESSABLE_ENTITY
+    assert error in response.json()["detail"]
 
 
 @pytest.mark.django_db
@@ -95,11 +192,12 @@ def test_exchange_view_with_only_unknown_audience(source_api_client, caplog):
         "audience": "foo",
     }
     with caplog.at_level(logging.INFO):
-        response = source_api_client.post(
-            "/auth/token/exchange/", payload, content_type="application/json"
-        )
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.json() == {"detail": "Only unknown audience(s) requested."}
+        response = source_api_client.post("/exchange/", data=payload)
+    assert response.status_code == 400  # BAD_REQUEST
+    assert response.json() == {
+        "detail": "Only unknown audience(s) requested.",
+        "code": "invalid_target",
+    }
     assert "Only unknown audience(s) requested: foo" in caplog.messages
 
 
@@ -114,11 +212,9 @@ def test_exchange_view_with_unknown_audience(source_api_client, caplog):
         "audience": "foo service:target",
     }
     with caplog.at_level(logging.INFO):
-        response = source_api_client.post(
-            "/auth/token/exchange/", payload, content_type="application/json"
-        )
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.json() == {"detail": "Unknown audience(s) requested."}
+        response = source_api_client.post("/exchange/", data=payload)
+    assert response.status_code == 400  # BAD_REQUEST
+    assert response.json() == {"detail": "Unknown audience(s) requested.", "code": "invalid_target"}
     assert "Unknown audience(s) requested: foo" in caplog.messages
 
 
@@ -137,11 +233,12 @@ def test_exchange_view_with_introspection_error(source_api_client, monkeypatch, 
         f"{settings.OIDC_RS_BACKEND_CLASS}.get_user_info_with_introspection",
         mock.Mock(side_effect=HTTPError("Connection fails")),
     )
-    response = source_api_client.post(
-        "/auth/token/exchange/", payload, content_type="application/json"
-    )
-    assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
-    assert response.json() == {"detail": "Failed to introspect subject token."}
+    response = source_api_client.post("/exchange/", data=payload)
+    assert response.status_code == 422  # UNPROCESSABLE_ENTITY
+    assert response.json() == {
+        "detail": "Failed to introspect subject token.",
+        "code": "invalid_introspection",
+    }
 
 
 @pytest.mark.django_db
@@ -162,11 +259,9 @@ def test_exchange_view_with_inactive_rule(source_api_client, source_service, cap
         "audience": "service:target service:other",
     }
     with caplog.at_level(logging.INFO):
-        response = source_api_client.post(
-            "/auth/token/exchange/", payload, content_type="application/json"
-        )
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.json() == {"detail": "Unknown audience(s) requested."}
+        response = source_api_client.post("/exchange/", data=payload)
+    assert response.status_code == 400  # BAD_REQUEST
+    assert response.json() == {"detail": "Unknown audience(s) requested.", "code": "invalid_target"}
     assert "Unknown audience(s) requested: service:other" in caplog.messages
 
 
@@ -190,10 +285,9 @@ def test_exchange_view_with_subject_token_type(
         "audience": "service:target",
     }
     # Create an exchange token
-    response = source_api_client.post(
-        "/auth/token/exchange/", payload, content_type="application/json"
-    )
-    assert response.status_code == status.HTTP_200_OK
+    # response = source_api_client.post("/exchange/", payload)
+    response = source_api_client.post("/exchange/", data=payload)
+    assert response.status_code == 200  # OK
     exchanged_token = response.json()
     assert "access_token" in exchanged_token
     assert len(exchanged_token["access_token"]) > 1
@@ -230,10 +324,8 @@ def test_exchange_view_with_requested_token_type(requested_token_type, source_ap
         "requested_token_type": requested_token_type,
     }
     # Create an exchange token
-    response = source_api_client.post(
-        "/auth/token/exchange/", payload, content_type="application/json"
-    )
-    assert response.status_code == status.HTTP_200_OK
+    response = source_api_client.post("/exchange/", data=payload)
+    assert response.status_code == 200  # OK
     exchanged_token = response.json()
     assert "access_token" in exchanged_token
     assert len(exchanged_token["access_token"]) > 1
@@ -266,45 +358,90 @@ def test_exchange_view_with_requested_refresh_token_type(source_api_client):
         "requested_token_type": "urn:ietf:params:oauth:token-type:refresh_token",
     }
     # Create an exchange token
-    response = source_api_client.post(
-        "/auth/token/exchange/", payload, content_type="application/json"
-    )
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
+    response = source_api_client.post("/exchange/", data=payload)
+    assert response.status_code == 422  # UNPROCESSABLE_ENTITY
     payload = response.json()
-    assert payload["error"] == "invalid_request"
-    assert (
-        "Invalid enum value 'urn:ietf:params:oauth:token-type:refresh_token'"
-        " - at `$.requested_token_type`"
-    ) in payload["error_description"]
+    assert {
+        "type": "enum",
+        "loc": ["form", "requested_token_type"],
+        "msg": (
+            "Input should be 'urn:ietf:params:oauth:token-type:access_token' or "
+            "'urn:ietf:params:oauth:token-type:jwt'"
+        ),
+        "ctx": {
+            "expected": (
+                "'urn:ietf:params:oauth:token-type:access_token' or "
+                "'urn:ietf:params:oauth:token-type:jwt'"
+            )
+        },
+    } in payload["detail"]
 
 
 @pytest.mark.django_db
-def test_instrospect_view_auth():
-    """Test the TokenIntrospectView authentication."""
-    client = APIClient()
-    response = client.post("/auth/token/introspect/", {}, content_type="application/json")
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize("payload", [{}, {"token": None}, {"token": ""}])
-def test_introspect_view_without_token(target_api_client, payload):
-    """Test the TokenIntrospectView with no or empty exchange token."""
-    response = target_api_client.post(
-        "/auth/token/introspect/", payload, content_type="application/json"
-    )
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == {"active": False}
+@pytest.mark.parametrize(
+    ("payload", "error"),
+    [
+        (
+            {},
+            {
+                "type": "missing",
+                "loc": [
+                    "form",
+                    "token",
+                ],
+                "msg": "Field required",
+            },
+        ),
+        (
+            {"token": None},
+            {
+                "type": "missing",
+                "loc": [
+                    "form",
+                    "token",
+                ],
+                "msg": "Field required",
+            },
+        ),
+        (
+            {"token": ""},
+            {
+                "type": "string_pattern_mismatch",
+                "loc": [
+                    "form",
+                    "token",
+                ],
+                "msg": "String should match pattern '^\\s*\\S{32,}\\s*$'",
+                "ctx": {"pattern": "^\\s*\\S{32,}\\s*$"},
+            },
+        ),
+        (
+            {"token": "0123456789"},
+            {
+                "type": "string_pattern_mismatch",
+                "loc": [
+                    "form",
+                    "token",
+                ],
+                "msg": "String should match pattern '^\\s*\\S{32,}\\s*$'",
+                "ctx": {"pattern": "^\\s*\\S{32,}\\s*$"},
+            },
+        ),
+    ],
+)
+def test_introspect_view_without_token(target_api_client, payload, error):
+    """Test the TokenIntrospectView with no, empty or too short exchange token."""
+    response = target_api_client.post("/introspect/", data=payload)
+    assert response.status_code == 422  # UNPROCESSABLE_ENTITY
+    assert error in response.json()["detail"]
 
 
 @pytest.mark.django_db
 def test_introspect_view_with_unknown_token(target_api_client, caplog):
     """Test the TokenIntrospectView with no or empty exchange token."""
     with caplog.at_level(logging.INFO):
-        response = target_api_client.post(
-            "/auth/token/introspect/", {"token": uuid4().hex}, content_type="application/json"
-        )
-    assert response.status_code == status.HTTP_200_OK
+        response = target_api_client.post("/introspect/", data={"token": uuid4().hex})
+    assert response.status_code == 200  # OK
     assert response.json() == {"active": False}
     assert "Token not found." in caplog.messages
 
@@ -339,11 +476,10 @@ def test_introspect_view_invalid_token(  # noqa: PLR0913
     )
     with caplog.at_level(logging.INFO):
         response = target_api_client.post(
-            "/auth/token/introspect/",
-            {"token": exchanged_token.token},
-            content_type="application/json",
+            "/introspect/",
+            data={"token": exchanged_token.token},
         )
-    assert response.status_code == status.HTTP_200_OK
+    assert response.status_code == 200  # OK
     assert response.json() == {"active": False}
     assert (
         f"Token is invalid: token_jti={exchanged_token.subject_token_jti}, "
@@ -362,11 +498,10 @@ def test_introspect_view_invalid_jwt_signature(target_api_client, target_service
         mock.patch.object(TokenGenerator, "verify_jwt", side_effect=ValueError("wrong signature")),
     ):
         response = target_api_client.post(
-            "/auth/token/introspect/",
-            {"token": exchanged_token.token},
-            content_type="application/json",
+            "/introspect/",
+            data={"token": exchanged_token.token},
         )
-    assert response.status_code == status.HTTP_200_OK
+    assert response.status_code == 200  # OK
     assert response.json() == {"active": False}
     assert "JWT signature verification failed (wrong signature)" in caplog.messages
 
@@ -378,12 +513,8 @@ def test_introspect_view_invalid_audience(target_api_client, caplog):
         token_type=TokenTypeChoices.JWT, audiences=["service:foo"]
     )
     with caplog.at_level(logging.INFO):
-        response = target_api_client.post(
-            "/auth/token/introspect/",
-            {"token": exchanged_token.token},
-            content_type="application/json",
-        )
-    assert response.status_code == status.HTTP_200_OK
+        response = target_api_client.post("/introspect/", data={"token": exchanged_token.token})
+    assert response.status_code == 200  # OK
     assert response.json() == {"active": False}
     assert (
         "'service:target' service tried to act on an exchanged token that is beyond its audience"
@@ -401,13 +532,13 @@ def test_introspect_view_introspection_with_wrong_token_type_hint(
     )
     with caplog.at_level(logging.INFO):
         response = target_api_client.post(
-            "/auth/token/introspect/",
-            {"token": exchanged_token.token, "token_type_hint": token_type_hint},
-            content_type="application/json",
+            "/introspect/",
+            data={"token": exchanged_token.token, "token_type_hint": token_type_hint},
         )
-    assert response.status_code == status.HTTP_200_OK
+    assert response.status_code == 200  # OK
     introspected_token = response.json()
-    assert not introspected_token["active"]
+    # the token exists  and is active (token type hint should be ignored)
+    assert introspected_token["active"]
 
 
 @pytest.mark.django_db
@@ -420,11 +551,10 @@ def test_introspect_view_introspection_with_token_type_hint(
     )
     with caplog.at_level(logging.INFO):
         response = target_api_client.post(
-            "/auth/token/introspect/",
-            {"token": exchanged_token.token, "token_type_hint": "access_token"},
-            content_type="application/json",
+            "/introspect/",
+            data={"token": exchanged_token.token, "token_type_hint": "access_token"},
         )
-    assert response.status_code == status.HTTP_200_OK
+    assert response.status_code == 200  # OK
     introspected_token = response.json()
     assert introspected_token["active"]
     assert (
@@ -442,11 +572,10 @@ def test_introspect_view_introspection(target_api_client, target_service, caplog
     )
     with caplog.at_level(logging.INFO):
         response = target_api_client.post(
-            "/auth/token/introspect/",
-            {"token": exchanged_token.token},
-            content_type="application/json",
+            "/introspect/",
+            data={"token": exchanged_token.token},
         )
-    assert response.status_code == status.HTTP_200_OK
+    assert response.status_code == 200  # OK
     assert (
         f"Token introspected (active): token_jti={exchanged_token.subject_token_jti}, "
         f"type={exchanged_token.token_type}, kid={exchanged_token.jwt_kid or 'N/A'}"
@@ -457,32 +586,70 @@ def test_introspect_view_introspection(target_api_client, target_service, caplog
 
 
 @pytest.mark.django_db
-def test_revocation_view_auth():
-    """Test the TokenRevocationView authentication."""
-    client = APIClient()
-    response = client.post("/auth/token/revoke/", {}, content_type="application/json")
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize("payload", [{}, {"token": None}, {"token": ""}])
-def test_revocation_view_without_token(source_api_client, payload):
+@pytest.mark.parametrize(
+    ("payload", "error"),
+    [
+        (
+            {},
+            {
+                "type": "missing",
+                "loc": [
+                    "form",
+                    "token",
+                ],
+                "msg": "Field required",
+            },
+        ),
+        (
+            {"token": None},
+            {
+                "type": "missing",
+                "loc": [
+                    "form",
+                    "token",
+                ],
+                "msg": "Field required",
+            },
+        ),
+        (
+            {"token": ""},
+            {
+                "type": "string_pattern_mismatch",
+                "loc": [
+                    "form",
+                    "token",
+                ],
+                "msg": "String should match pattern '^\\s*\\S{32,}\\s*$'",
+                "ctx": {"pattern": "^\\s*\\S{32,}\\s*$"},
+            },
+        ),
+        (
+            {"token": "0123456789"},
+            {
+                "type": "string_pattern_mismatch",
+                "loc": [
+                    "form",
+                    "token",
+                ],
+                "msg": "String should match pattern '^\\s*\\S{32,}\\s*$'",
+                "ctx": {"pattern": "^\\s*\\S{32,}\\s*$"},
+            },
+        ),
+    ],
+)
+def test_revocation_view_without_token(source_api_client, payload, error):
     """Test the TokenRevocationView instrospection with no or empty token."""
-    response = source_api_client.post(
-        "/auth/token/revoke/", payload, content_type="application/json"
-    )
-    assert response.status_code == status.HTTP_200_OK
-    assert not len(response.content)
+    response = source_api_client.post("/revoke/", data=payload)
+    assert response.status_code == 422  # UNPROCESSABLE_ENTITY
+    assert error in response.json()["detail"]
 
 
 @pytest.mark.django_db
 def test_revocation_view_with_unknown_token(source_api_client, caplog):
     """Test the TokenRevocationView instrospection with no or empty token."""
     with caplog.at_level(logging.INFO):
-        response = source_api_client.post(
-            "/auth/token/revoke/", {"token": uuid4().hex}, content_type="application/json"
-        )
-    assert response.status_code == status.HTTP_200_OK
+        response = source_api_client.post("/revoke/", data={"token": uuid4().hex})
+    assert response.status_code == 200  # OK
     assert not len(response.content)
     assert "Token revocation failed (not found)." in caplog.messages
 
@@ -500,10 +667,8 @@ def test_revocation_view(target_api_client, target_service, caplog):
 
     before = datetime.now(tz=UTC)
     with caplog.at_level(logging.INFO):
-        response = target_api_client.post(
-            "/auth/token/revoke/", {"token": exchanged_token.token}, content_type="application/json"
-        )
-    assert response.status_code == status.HTTP_200_OK
+        response = target_api_client.post("/revoke/", data={"token": exchanged_token.token})
+    assert response.status_code == 200  # OK
     assert not len(response.content)
     assert (
         "Token revoked: "